Introduce std::array as Grid type
std::array has several advantages over C arrays, especially the length safety.
C arrays, when passed to a function, decay to pointers and we have to make sure
we get the sizes right when iterating over them.
in[N][N] as function parameter is basically a lie.

std::array is also easy to deal with in terms of all the usual standard container
functions, e.g. begin()/end(), size(), etc.