Find unassigned locations once
Instead of iterating over the beginning of the grid every time we need the next location, we can find them all before we start solving the Sudoku.

Also, references for out parameters are quite ugly and unintuitive, so they're best avoided.

Introduces a new type for grid locations and uses size_t instead of int for coordinates and grid sizes.